src/test/java/com/leetcode/strings/medium/VowelSpellchecker_966Test.java
package com.leetcode.strings.medium;

import org.junit.jupiter.api.extension.ExtensionContext;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.ArgumentsProvider;
import org.junit.jupiter.params.provider.ArgumentsSource;

import java.util.stream.Stream;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.params.provider.Arguments.arguments;

class VowelSpellchecker_966Test {
  private final VowelSpellchecker_966 solution = new VowelSpellchecker_966();

  private static class TestArgumentsProvider implements ArgumentsProvider {
    @Override
    public Stream<? extends Arguments> provideArguments(ExtensionContext context) {
      return Stream.of(
          arguments(
              new String[] {"KiTe", "kite", "hare", "Hare"},
              new String[] {
                "kite", "Kite", "KiTe", "Hare", "HARE", "Hear", "hear", "keti", "keet", "keto"
              },
              new String[] {"kite", "KiTe", "KiTe", "Hare", "hare", "", "", "KiTe", "", "KiTe"}));
    }
  }

  @ParameterizedTest
  @ArgumentsSource(TestArgumentsProvider.class)
  void spellchecker(String[] wordlist, String[] queries, String[] expectedResult) {
    String[] actualResult = solution.spellchecker(wordlist, queries);

    assertThat(actualResult).isEqualTo(expectedResult);
  }
}
